requirements.txt
cachetools==4.2.2
google-api-core==1.31.1
google-api-python-client==2.15.0
google-auth==1.34.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.5
googleapis-common-protos==1.53.0
httplib2==0.19.1
oauthlib==3.1.1
packaging==21.0
protobuf==3.17.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==2.4.7
requests-oauthlib==1.3.0
rsa==4.7.2
uritemplate==3.0.1
certifi==2021.5.30
charset-normalizer==2.0.4
click==8.0.1
Flask==2.0.1
gunicorn==20.1.0
idna==3.2
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
PyJWT==1.7.1
PySocks==1.7.1
pytz==2021.1
requests==2.26.0
six==1.16.0
twilio==6.63.0
urllib3==1.26.6
Werkzeug==2.0.1
APScheduler==3.7.0
tzlocal==2.1
